requirements.txt
cffi==1.1.2
cryptography==0.9.1
enum34==1.0.4
idna==2.0
ipaddress==1.0.7
ndg-httpsclient==0.4.0
pyasn1==0.1.7
pycparser==2.14
pyOpenSSL==0.15.1
python-dateutil==2.4.2
requests==2.7.0
six==1.9.0
